refactor: create a custom hook
