+++ src/pages/_app.js
@@ -1,5 +1,4 @@
 import { AuthProvider } from '@/contexts/authContext';
-import { DocumentationEditorProvider } from '@/contexts/documentationEditorContext';
 import '@/styles/globals.scss';
 
 import Head from 'next/head';
@@ -11,9 +10,7 @@ export default function App({ Component, pageProps }) {
         <meta name="viewport" content="width=device-width, initial-scale=1" />
       </Head>
       <AuthProvider>
-        <DocumentationEditorProvider>
-          <Component {...pageProps} />
-        </DocumentationEditorProvider>
+        <Component {...pageProps} />
       </AuthProvider>
     </>
   );
